Kommentare ergänzt und Code Style verbessert
